// apps/frontend/src/api/generated.ts
export interface components {
    schemas: {
// ...
        CitationDTO: {
            /** Citation Id */
            citation_id: string;
            /** Document Title */
            document_title: string;
            /** Evidence Id */
            evidence_id: string;
            /** Pdf Pages */
            pdf_pages: number[];
            /** Printed Page Labels */
            printed_page_labels: string[];
            /** Source Url */
            source_url: string;
            /** Structural Path */
            structural_path: string[];
            /** Version Label */
            version_label?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}